net: Moved checksum stuff out from common to separate header
